sonar dedupe code
